refatorado client feing para processar Page

+++ ds-produtos/src/main/java/com/farias/banco/dsprodutos/dto/ProdutoValorDTO.java
@@ -11,6 +11,6 @@ import lombok.With;
 @With
 public class ProdutoValorDTO {
 
-	private Integer produto;
+	private Long produto;
 	private BigDecimal valor;
 }
