+++ ds-produtos/src/main/java/com/farias/banco/dsprodutos/repository/ProdutosFaixaRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import com.farias.banco.dsprodutos.model.ProdutoFaixa;
 
 @Repository
-public interface ProdutosFaixaRepository extends JpaRepository<ProdutoFaixa, Integer>, JpaSpecificationExecutor<ProdutoFaixa> {
+public interface ProdutosFaixaRepository extends JpaRepository<ProdutoFaixa, Long>, JpaSpecificationExecutor<ProdutoFaixa> {
 
 	@Query(value = "select * from produtofaixa where :score between prof_scoremin and prof_scoremax", nativeQuery = true)
 	List<ProdutoFaixa> findByScoreFaixa(@Param("score") Integer score);
